+++ cpp/subprojects/boosting/include/boosting/output/predictor_probability_auto.hpp
@@ -1,0 +1,50 @@
+/*
+ * @author Michael Rapp (michael.rapp.ml@gmail.com)
+ */
+#pragma once
+
+#include "common/output/predictor_probability.hpp"
+#include "common/multi_threading/multi_threading.hpp"
+#include "boosting/losses/loss.hpp"
+
+
+namespace boosting {
+
+    /**
+     * Allows to configure a predictor that automatically decides for a method that is used to predict probabilities for
+     * given query examples, which estimate the chance of individual labels to be relevant.
+     */
+    class AutomaticProbabilityPredictorConfig final : public IProbabilityPredictorConfig {
+
+        private:
+
+            const std::unique_ptr<ILossConfig>& lossConfigPtr_;
+
+            const std::unique_ptr<IMultiThreadingConfig>& multiThreadingConfigPtr_;
+
+        public:
+
+            /**
+             * @param lossConfigPtr             A reference to an unique pointer that stores the configuration of the
+             *                                  loss function
+             * @param multiThreadingConfigPtr   A reference to an unique pointer that stores the configuration of the
+             *                                  multi-threading behavior that should be used to predict for several
+             *                                  query examples in parallel
+             */
+            AutomaticProbabilityPredictorConfig(const std::unique_ptr<ILossConfig>& lossConfigPtr,
+                                                const std::unique_ptr<IMultiThreadingConfig>& multiThreadingConfigPtr);
+
+            /**
+             * @see `IProbabilityPredictorConfig::createProbabilityPredictorFactory`
+             */
+            std::unique_ptr<IProbabilityPredictorFactory> createProbabilityPredictorFactory(
+                const IFeatureMatrix& featureMatrix, uint32 numLabels) const override;
+
+            /**
+             * @see `IPredictorConfig::isLabelVectorSetNeeded`
+             */
+            bool isLabelVectorSetNeeded() const override;
+
+    };
+
+}