+++ cpp/subprojects/boosting/src/boosting/learner.cpp
@@ -13,7 +13,9 @@
 #include "boosting/output/predictor_classification_example_wise.hpp"
 #include "boosting/output/predictor_classification_label_wise.hpp"
 #include "boosting/output/predictor_regression_label_wise.hpp"
+#include "boosting/output/predictor_probability_auto.hpp"
 #include "boosting/output/predictor_probability_label_wise.hpp"
+#include "boosting/output/predictor_probability_marginalized.hpp"
 #include "boosting/rule_evaluation/head_type_auto.hpp"
 #include "boosting/rule_evaluation/head_type_complete.hpp"
 #include "boosting/rule_evaluation/head_type_single.hpp"
@@ -38,7 +40,7 @@ namespace boosting {
         this->useAutomaticLabelBinning();
         this->useAutomaticClassificationPredictor();
         this->useLabelWiseRegressionPredictor();
-        this->useLabelWiseProbabilityPredictor();
+        this->useAutomaticProbabilityPredictor();
     }
 
     const IHeadConfig& BoostingRuleLearner::Config::getHeadConfig() const {
@@ -205,6 +207,16 @@ namespace boosting {
     void BoostingRuleLearner::Config::useLabelWiseProbabilityPredictor() {
         probabilityPredictorConfigPtr_ =
             std::make_unique<LabelWiseProbabilityPredictorConfig>(lossConfigPtr_, parallelPredictionConfigPtr_);
+    }
+
+    void BoostingRuleLearner::Config::useMarginalizedProbabilityPredictor() {
+        probabilityPredictorConfigPtr_ =
+            std::make_unique<MarginalizedProbabilityPredictorConfig>(lossConfigPtr_, parallelPredictionConfigPtr_);
+    }
+
+    void BoostingRuleLearner::Config::useAutomaticProbabilityPredictor() {
+        probabilityPredictorConfigPtr_ =
+            std::make_unique<AutomaticProbabilityPredictorConfig>(lossConfigPtr_, parallelPredictionConfigPtr_);
     }
 
     BoostingRuleLearner::BoostingRuleLearner(std::unique_ptr<IBoostingRuleLearner::IConfig> configPtr,