+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 A major update to the BOOMER algorithm that introduces the following changes:
 
 * The parameter `--head-type` does now allow to learn partial rules that predict for a predefined number of labels (`partial-fixed`).
+* A new parameter `--probability-predictor` has been added. It allows to specify whether probability estimates should be obtained for each label independently (`label-wise`) or via marginization over the label vectors that are encountered in the training data (`marginalized`).
 
 ### Version 0.8.1 (Mar. 4th, 2022)